refactor: move stats_table_row_cache out
