+++ pkg/statistics/handle/cache/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "cache",
     srcs = [
+        "stats_table_row_cache.go",
         "statscache.go",
         "statscacheinner.go",
     ],
